requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
dataclasses-json==0.5.7
dnspython==2.3.0
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
imageio-ffmpeg==0.4.8
langchain==0.0.154
loguru==0.7.0
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.1
packaging==23.1
pinecone-client==2.2.1
protobuf==4.22.3
py-cord==2.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycord==0.1.1
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydub==0.25.1
PyNaCl==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.12
tenacity==8.2.2
tiktoken==0.3.0
tqdm==4.64.1
typing-extensions==4.5.0
typing-inspect==0.8.0
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
